refactor: move createTableSettings to its own file
