+++ packages/components/src/DataTable/DataTable.tsx
@@ -1,17 +1,9 @@
-import React, { Dispatch, SetStateAction } from "react";
-import {
-  ColumnDef,
-  PaginationState,
-  SortingState,
-  TableOptions,
-  flexRender,
-  getCoreRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-import { Footer } from "./Footer";
+import { ColumnDef, flexRender, useReactTable } from "@tanstack/react-table";
+import React from "react";
+import { createTableSettings } from "./createTableSettings";
 import styles from "./DataTable.css";
+import { Footer } from "./Footer";
+import { Pagination, Sorting } from "./types";
 import { Icon } from "../Icon";
 
 interface DataTableProps<T> {
@@ -49,75 +41,6 @@ interface DataTableProps<T> {
   height?: number;
 
   stickyHeader?: boolean;
-}
-
-interface Pagination {
-  /**
-   * When manually controlled represents the current pagination state of the table.
-   *
-   * @type {PaginationState}
-   */
-  state?: PaginationState;
-
-  /**
-   * When manually controlling pagination, you should supply a total pageCount value to the table if you know it.
-   * If you do not know how many pages there are, you can set this to -1. (Or should we error?)
-   *
-   * @type {number}
-   */
-  pageCount?: number;
-
-  /**
-   * If this function is provided, it will be called when the pagination state changes and you will be expected
-   * to manage the state yourself. You can pass the managed state back to the table via the state option.
-   */
-  onPaginationChange?: Dispatch<SetStateAction<PaginationState>>;
-
-  /**
-   * Enables manual pagination. If this option is set to true,
-   * the table will not automatically paginate and instead will expect you
-   * to manually paginate the rows before passing them to the table.
-   *
-   * @type {boolean}
-   */
-  manualPagination: boolean;
-  /**
-   * The options to control the pagination
-   *
-   * @type {number[]}
-   * @memberof Pagination
-   */
-  itemsPerPage?: number[];
-  /**
-   * When manual pagination is enable the total numbers of
-   * records in the table should be provided.
-   *
-   * @type {number}
-   * @memberof Pagination
-   */
-  totalItems?: number;
-}
-
-interface Sorting {
-  /**
-   * Represents the current sorting state of the table.
-   *
-   * @type {SortingState}
-   */
-  state?: SortingState;
-
-  /**
-   * Enables manual sorting for the table.
-   *
-   * @type {boolean}
-   */
-  manualSorting: boolean;
-
-  /**
-   * This overrides the default internal state management,
-   * so you will need to persist the state change either fully or partially outside of the table.
-   */
-  onSortingChange?: Dispatch<SetStateAction<SortingState>>;
 }
 
 export function DataTable<T extends object>({
@@ -214,90 +137,4 @@ export function DataTable<T extends object>({
       )}
     </div>
   );
-}
-
-function createTableSettings<T>(
-  data: T[],
-  columns: ColumnDef<T>[],
-  options?: { pagination?: Pagination; sorting?: Sorting },
-) {
-  const { state: paginationState, ...restPaginationSettings } =
-    getPaginationSettings(options?.pagination);
-
-  const { state: sortingState, ...restSortingSettings } = getSortingSettings(
-    options?.sorting,
-  );
-
-  const tableSettings: TableOptions<T> = {
-    data,
-    columns,
-    state: { ...paginationState, ...sortingState },
-    getCoreRowModel: getCoreRowModel(),
-    ...restPaginationSettings,
-    ...restSortingSettings,
-  };
-
-  return tableSettings;
-}
-
-type PaginationSettings<T> = Pick<
-  TableOptions<T>,
-  | "state"
-  | "pageCount"
-  | "onPaginationChange"
-  | "manualPagination"
-  | "getPaginationRowModel"
->;
-
-function getPaginationSettings<T>(
-  pagination?: Pagination,
-): PaginationSettings<T> {
-  const paginationSettings: PaginationSettings<T> = {};
-
-  if (!pagination) {
-    return paginationSettings;
-  }
-
-  const { manualPagination, state, pageCount, onPaginationChange } = pagination;
-
-  paginationSettings.manualPagination = manualPagination;
-
-  if (manualPagination) {
-    if (state) paginationSettings.state = { pagination: state };
-
-    if (pageCount) paginationSettings.pageCount = pageCount;
-
-    if (onPaginationChange) {
-      paginationSettings.onPaginationChange = onPaginationChange;
-    }
-  } else {
-    paginationSettings.getPaginationRowModel = getPaginationRowModel();
-  }
-
-  return paginationSettings;
-}
-
-type SortingSettings<T> = Pick<
-  TableOptions<T>,
-  "state" | "manualSorting" | "onSortingChange" | "getSortedRowModel"
->;
-
-function getSortingSettings<T>(sorting?: Sorting): SortingSettings<T> {
-  const sortingSettings: SortingSettings<T> = {};
-
-  if (!sorting) return sortingSettings;
-
-  const { manualSorting, onSortingChange, state } = sorting;
-
-  sortingSettings.manualSorting = manualSorting;
-
-  if (manualSorting) {
-    if (state) sortingSettings.state = { sorting: state };
-
-    if (onSortingChange) sortingSettings.onSortingChange = onSortingChange;
-  } else {
-    sortingSettings.getSortedRowModel = getSortedRowModel();
-  }
-
-  return sortingSettings;
 }
